requirements.txt
Django==4.0.3
aiogram==2.19
beautifulsoup4==4.11.1
djangorestframework==3.13.1
requests==2.27.1
psycopg2-binary==2.9.5
lxml==4.8.0
gunicorn==20.1.0
django-cors-headers==3.13.0